Enemy react to Player's current move
Enemy now reacts, and gives the appearance of reacting to the players
current move.  This makes the game even more difficult, and Enemy's
present behaviour will probably eventually become the prototype
for a `nightmare' or `hard' mode.

Once Enemy has caught of up to Player, the screen now locks and informs
the player:
    >  GAME OVER!
    >  Press `q' to quit.

* Remove temp variables ex, ey (Enemy pos).
* Move Enemy's seek behaviour to _AFTER_ Player has made her move.

* While here:
  * Remove TODO note.
  * Remove commented out info panel text - can always reach into
    git history and add this back for debugging.
  * Add space after `info:' label to info message.
  * mark parm in Actor::atop as const - good practice.  Should
    probably be doing more of this.