Add more enemies to game area
Now that enemies can be defeated by the player, we add more for fun
and challenge.

To facilitate this, we refactor the player-enemy distance checker to
run significantly faster. To whit: rather than re-initialising all
enemies if only __one__ is too close, we simply reroll each enemy's
starting position until they are the desired distance from the player.