[#60] Implement Actor Action points
Finishes #60

Implements simple action points (moves per global tick (turn)) for
enemies.  For present purposes, stamina will suffice for moving and
fighting, but actions (such as free actions) may be added later, so
this is left in as a _stub_.

Also adds a list of "messages" to actors - populated when an actor
either attacks or kills another actor.  Eventually, we will want a
method which may retrieve __all__ new messages since the last one
was retrieved, but for now the very last message suffices.

* Adds a default name to all enemies.
* Adds a method to retrieve a list of items in Player's inventory.

* While here:
  * Remove temp variable `ch,' and instead call the game windows'
    getch directly from the switch.
  * Add missing break after ESC key in said switch.